refactored things into their own packages
