+++ andhow/src/main/java/yarnandtail/andhow/sample/BaseSamplePrinter.java
@@ -3,15 +3,14 @@
 package yarnandtail.andhow.sample;
 
 import java.io.PrintStream;
-import java.util.ArrayList;
 import java.util.List;
 import yarnandtail.andhow.GroupInfo;
 import yarnandtail.andhow.Property;
 import yarnandtail.andhow.PropertyGroup;
-import static yarnandtail.andhow.util.ReportGenerator.DEFAULT_LINE_WIDTH;
 import yarnandtail.andhow.util.TextUtil;
 import yarnandtail.andhow.Validator;
 import yarnandtail.andhow.SamplePrinter;
+import yarnandtail.andhow.sample.TextLine.HRLine;
 
 /**
  *
@@ -19,7 +18,19 @@ import yarnandtail.andhow.SamplePrinter;
  */
 public abstract class BaseSamplePrinter implements SamplePrinter {
 	
-	//private static final String LINE_PREFIX = "# ";
+	abstract PrintFormat getFormat();
+	
+	public abstract TextBlock getSampleFileStart();
+	
+	/**
+	 * Must provide its own line comment prefix if the format has no block comment.
+	 * @return 
+	 */
+	public abstract TextBlock getSampleStartComment();
+	
+	public abstract TextBlock getActualProperty(Class<? extends PropertyGroup> group, Property prop) throws Exception;
+	
+	public abstract String getSampleFileEnd();
 	
 	protected void printIfNotNull(PrintStream out, String line) {
 		if (line != null) {
@@ -27,15 +38,16 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 		}
 	}
 	
-	abstract Format getFormat();
-	
-	protected void print(PrintStream out, TextBlock block, Format format) {
+	protected void print(PrintStream out, TextBlock block, PrintFormat format) {
 		
-		if (block == null || block.getLines().size() == 0) return;
+		if (block == null) return;
 		
 		boolean useBlockComment = false;
 		boolean useLineComment = false;
 		
+		if (block.isBlankLineBefore()) {
+			out.println();
+		}
 		
 		if (block.isComment()) {
 			useLineComment = format.usesLineComments();
@@ -46,9 +58,9 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 			
 			if (block.getLines().size() == 1) {
 				
-				Line line = block.getLine(0);
+				TextLine line = block.getLine(0);
 				
-				if (block.wrap) {
+				if (block.isWrap()) {
 					List<String> lines = line.getWrappedBlockComment(format, true, true);
 					lines.stream().forEachOrdered(l -> out.println(l));
 				} else {
@@ -62,7 +74,7 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 				
 				while (rowToPrint < block.getLines().size()) {
 
-					Line line = block.getLine(rowToPrint);
+					TextLine line = block.getLine(rowToPrint);
 					boolean isFirstLine = rowToPrint == 0;
 					boolean isLastLine = (rowToPrint == (block.getLines().size() - 1));
 					
@@ -82,7 +94,7 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 			
 		} else if (useLineComment) {
 			
-			for (Line line : block.getLines()) {
+			for (TextLine line : block.getLines()) {
 				if (line.wrap) {
 					List<String> lines = line.getWrappedLineComment(format);
 					lines.stream().forEachOrdered(l -> out.println(l));
@@ -93,7 +105,7 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 						
 		} else {
 			
-			for (Line line : block.getLines()) {
+			for (TextLine line : block.getLines()) {
 				
 				if (line instanceof HRLine) {
 					if (format.usesBlockComments()) {
@@ -110,19 +122,11 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 			}
 			
 		}
+		
+		if (block.isBlankLineAfter()) {
+			out.println();
+		}
 	}
-	
-	public abstract TextBlock getSampleFileStart();
-	
-	/**
-	 * Must provide its own line comment prefix if the format has no block comment.
-	 * @return 
-	 */
-	public abstract TextBlock getSampleStartComment();
-	
-	public abstract TextBlock getActualProperty(Class<? extends PropertyGroup> group, Property prop) throws Exception;
-	
-	public abstract String getSampleFileEnd();
 
 	@Override
 	public void printSampleStart(PrintStream out) {
@@ -206,7 +210,11 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 		}
 		
 		print(out, tb, getFormat());
-		print(out, getActualProperty(group, prop), getFormat());
+		
+		TextBlock aptb = getActualProperty(group, prop);
+		aptb.setBlankLineAfter(true);	//Always want a space after
+		
+		print(out, aptb, getFormat());
 		
 	}
 	
@@ -221,318 +229,6 @@ public abstract class BaseSamplePrinter implements SamplePrinter {
 		out.println();
 		
 		printIfNotNull(out, getSampleFileEnd());
-	}
-	
-	public static class Format {
-		
-		/** either block comments or lineComment should be non-null, but not both */
-		public String blockCommentStart;
-		public String blockCommentEnd;
-		public String blockCommentSeparator = "\t"; //Use between the last text and the comment end if on same line
-		public String lineCommentPrefix;
-		public String lineCommentPrefixSeparator = " ";	//Separate the opening line comment from the text
-		public String secondLineIndent;
-		public String hr;
-		public int lineWidth;
-		
-		public boolean usesLineComments() {
-			return (TextUtil.trimToNull(lineCommentPrefix) != null);
-		}
-		
-		public boolean usesBlockComments() {
-			return ! usesLineComments();
-		}
-	}
-	
-	public abstract static class Line {
-		protected Boolean wrap;
-
-		abstract String getLine(Format format);
-		abstract String getLineComment(Format format);
-		abstract String getBlockComment(Format format, boolean startComment, boolean endComment);
-		
-		abstract List<String> getWrappedLine(Format format);
-		abstract List<String> getWrappedLineComment(Format format);
-		abstract List<String> getWrappedBlockComment(Format format, boolean startComment, boolean endComment);
-	}
-	
-	public static class StringLine extends Line {
-		String line;
-		public StringLine(String line, boolean wrap) {
-			this.line = line;
-			this.wrap = wrap;
-		}
-		
-		public StringLine(String line) {
-			this.line = line;
-			this.wrap = null;
-		}
-		
-		@Override
-		public String getLine(Format format) {
-			return line;
-		}
-		
-		@Override
-		public String getLineComment(Format format) {
-			return format.lineCommentPrefix + format.lineCommentPrefixSeparator + line;
-		}
-		
-		@Override
-		public String getBlockComment(Format format, boolean startComment, boolean endComment) {
-			
-			String out = line;
-
-				
-			if (startComment) {
-				out = format.blockCommentStart + format.blockCommentSeparator + out;
-			}
-
-			if (endComment) {
-				out = out + format.blockCommentSeparator + format.blockCommentEnd;
-			}
-			
-			return out;
-		}
-		
-		@Override
-		public List<String> getWrappedLine(Format format) {
-			return TextUtil.wrap(line, format.lineWidth, "", format.secondLineIndent);
-		}
-		
-		@Override
-		public List<String> getWrappedLineComment(Format format) {
-			return TextUtil.wrap(line, format.lineWidth, 
-					format.lineCommentPrefix + format.lineCommentPrefixSeparator, format.secondLineIndent);
-		}
-		
-		@Override
-		public List<String> getWrappedBlockComment(Format format, boolean startComment, boolean endComment) {
-			
-			List<String> lines = TextUtil.wrap(line, format.lineWidth, "", format.secondLineIndent);
-
-			if (startComment && endComment && lines.size() == 1) {
-				lines.set(0, format.blockCommentStart + format.blockCommentSeparator + 
-						lines.get(0) + format.blockCommentSeparator + format.blockCommentEnd);
-			} else {
-				if (startComment) {
-					lines.add(0, format.blockCommentStart);
-				}
-				
-				if (endComment) {
-					lines.set(lines.size() - 1, lines.get(lines.size() - 1) + format.blockCommentSeparator + format.blockCommentEnd);
-				}
-			}
-			
-			return lines;
-		}
-
-	}
-	
-	public static class HRLine extends Line {
-		public HRLine() {
-			wrap = false;
-		}
-		
-		@Override
-		public String getLine(Format format) {
-			return format.hr;
-		}
-		
-		@Override
-		public String getLineComment(Format format) {
-			return format.lineCommentPrefix + format.lineCommentPrefixSeparator + format.hr;
-		}
-		
-		@Override
-		public String getBlockComment(Format format, boolean startComment, boolean endComment) {
-			
-			String out = format.hr;
-
-				
-			if (startComment) {
-				out = format.blockCommentStart + format.blockCommentSeparator + out;
-			}
-
-			if (endComment) {
-				out = out + format.blockCommentSeparator + format.blockCommentEnd;
-			}
-			
-			return out;
-		}
-		
-		@Override
-		public List<String> getWrappedLine(Format format) {
-			ArrayList<String> lines = new ArrayList();
-			lines.add(format.hr);
-			return lines;
-		}
-		
-		@Override
-		public List<String> getWrappedLineComment(Format format) {
-			ArrayList<String> lines = new ArrayList();
-			lines.add(format.lineCommentPrefix + format.lineCommentPrefixSeparator + format.hr);
-			return lines;
-		}
-		
-		@Override
-		public List<String> getWrappedBlockComment(Format format, boolean startComment, boolean endComment) {
-			ArrayList<String> lines = new ArrayList();
-
-			if (startComment && endComment) {
-				lines.add(format.blockCommentStart + format.blockCommentSeparator + 
-						format.hr + format.blockCommentSeparator + format.blockCommentEnd);
-			} else {
-				if (startComment) {
-					lines.add(format.blockCommentStart + format.blockCommentSeparator + format.hr);
-				}
-				
-				if (endComment) {
-					lines.add(format.hr + format.blockCommentSeparator + format.blockCommentEnd);
-				}
-			}
-			
-			return lines;
-		}
-	}
-	
-
-	/**
-	 * A blank line will continue the comment or non-comment status of the block.
-	 * Use blankLineBefore/After of a TextBlock to get actual whitespace b/t
-	 * TextBlocks.
-	 */
-	public static class BlankLine extends Line {
-		public BlankLine() {
-			wrap = false;
-		}
-		
-		@Override
-		public String getLine(Format format) {
-			return "";
-		}
-		
-		@Override
-		public String getLineComment(Format format) {
-			return format.lineCommentPrefix + format.lineCommentPrefixSeparator + "";
-		}
-		
-		@Override
-		public String getBlockComment(Format format, boolean startComment, boolean endComment) {
-			
-			String out = "";
-
-				
-			if (startComment) {
-				out = format.blockCommentStart + format.blockCommentSeparator + out;
-			}
-
-			if (endComment) {
-				out = out + format.blockCommentSeparator + format.blockCommentEnd;
-			}
-			
-			return out;
-		}
-		
-		@Override
-		public List<String> getWrappedLine(Format format) {
-			ArrayList<String> lines = new ArrayList();
-			lines.add("");
-			return lines;
-		}
-		
-		@Override
-		public List<String> getWrappedLineComment(Format format) {
-			ArrayList<String> lines = new ArrayList();
-			lines.add(format.lineCommentPrefix + format.lineCommentPrefixSeparator + "");
-			return lines;
-		}
-		
-		@Override
-		public List<String> getWrappedBlockComment(Format format, boolean startComment, boolean endComment) {
-			ArrayList<String> lines = new ArrayList();
-
-			if (startComment && endComment) {
-				lines.add(format.blockCommentStart + format.blockCommentSeparator + 
-						"" + format.blockCommentSeparator + format.blockCommentEnd);
-			} else {
-				if (startComment) {
-					lines.add(format.blockCommentStart + format.blockCommentSeparator + "");
-				}
-				
-				if (endComment) {
-					lines.add("" + format.blockCommentSeparator + format.blockCommentEnd);
-				}
-			}
-			
-			return lines;
-		}
-	}
-	
-	public static class TextBlock {
-		boolean wrap;
-		boolean comment;
-		boolean blankLineBefore = false;
-		boolean blankLineAfter = false;
-		
-		ArrayList<Line> lines = new ArrayList();
-		
-		public TextBlock(boolean wrap, boolean comment) {
-			this.wrap = wrap;
-			this.comment = comment;
-		}
-		
-		public TextBlock(boolean wrap, boolean comment, boolean blankLineBefore, boolean blankLineAfter) {
-			this(wrap, comment);
-			this.blankLineBefore = blankLineBefore;
-			this.blankLineAfter = blankLineAfter;
-		}
-		
-		public TextBlock addLine(Line line) {
-			lines.add(line);
-			
-			if (line.wrap == null) {
-				line.wrap = wrap;
-			}
-			
-			return this;
-		}
-		
-		public TextBlock addLine(String strLine) {
-			StringLine line = new StringLine(strLine, wrap);
-			lines.add(line);
-			return this;
-		}
-		
-		public TextBlock addHR() {
-			lines.add(new HRLine());
-			return this;
-		}
-		
-		public TextBlock addBlank() {
-			lines.add(new BlankLine());
-			return this;
-		}
-
-		public boolean isWrap() {
-			return wrap;
-		}
-
-		public boolean isComment() {
-			return comment;
-		}
-
-		public ArrayList<Line> getLines() {
-			return lines;
-		}
-		
-		public Line getLine(int i) {
-			return lines.get(i);
-		}
-		
-		
-		
-		
 	}
 	
 }
